requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# OE2022_WC/app.py: 6
Pillow == 8.2.0

# OE2022_WC/app.py: 10
matplotlib == 3.4.2

# OE2022_WC/app.py: 8
nltk == 3.6.2

# OE2022_WC/app.py: 9
numpy == 1.20.2

# OE2022_WC/app.py: 12
scipy == 1.6.3

# OE2022_WC/app.py: 13
wordcloud == 1.8.1
